requirements.txt
streamlit
plotly
pdfplumber
faiss-cpu
sentence-transformers
numpy
groq
huggingface-hub
scikit-learn
torch